Restrict creation of arrays to use initial sizes
Using fixed-size arrays will better enforce the core CS concepts that this exercise is designed to teach.  The use of the append operator << allows a, dare I say, laziness that is opposed to the concept of a basic array in CS, which has implications not only historically and in lower-level languages, but also for performance and possible bug in codes (if accessing myArray[n+10] suddenly you have a bunch of nil values in between magically which may cause problems if unexpected.

I really like this exercise's purpose, but think that you should consider banning the use of << and requiring intentional declaration of arrays.